feat: customException으로 예외처리 구현, 게시글을 못찾는 경우

+++ src/main/java/minji/board/service/BoardService.java
@@ -2,6 +2,8 @@ package minji.board.service;
 
 import lombok.RequiredArgsConstructor;
 import minji.board.controller.dto.BoardRequestDTO;
+import minji.board.global.BoardException;
+import minji.board.global.BoardExceptionCode;
 import minji.board.model.BoardEntity;
 import minji.board.repository.BoardRepository;
 import org.springframework.data.domain.Sort;
@@ -24,7 +26,7 @@ public class BoardService {
 
     @Transactional(readOnly = true)
     public BoardEntity getBoard(Long boardId){
-        BoardEntity board = boardRepository.findById(boardId).orElseThrow(IllegalArgumentException::new);
+        BoardEntity board = boardRepository.findById(boardId).orElseThrow(()-> new BoardException(BoardExceptionCode.BOARD_NOT_FOUND));
         return board;
     }
 
@@ -36,14 +38,18 @@ public class BoardService {
 
     @Transactional
     public BoardEntity setBoard(Long boardId, BoardRequestDTO boardRequestDTO){
-        BoardEntity board = boardRepository.findById(boardId).orElseThrow(IllegalArgumentException::new);
+        BoardEntity board = boardRepository.findById(boardId).orElseThrow(()-> new BoardException(BoardExceptionCode.BOARD_NOT_FOUND));
         board.setBoard(boardRequestDTO.getTitle(), boardRequestDTO.getContent());
         BoardEntity responseBoard = boardRepository.save(board);
         return responseBoard;
     }
     @Transactional
     public String deleteBoard(Long boardId){
-        boardRepository.deleteById(boardId);
+        BoardEntity board = boardRepository.findById(boardId).orElseThrow(()-> new BoardException(BoardExceptionCode.BOARD_NOT_FOUND));
+        if(!board.getBoardId().equals(boardId)){
+            //삭제를 요청하는 게시글을 찾아서 없으면 에러 response 처리 후 하는 분기처리, 큰 의미는 없음
+            boardRepository.deleteById(boardId);
+        }
         return "DeleteBoard OK";
     }
